refactor: extract duplicated focus trap logic into useFocusTrap hook (#375)
Create a shared useFocusTrap hook that handles trigger element capture,
Tab/Shift+Tab cycling, focus restoration on unmount, and optional
Escape key callback. Refactor OverlayPanel, SidebarOverlay, and
FeedbackSheet to use the new hook, removing ~200 lines of duplicated
focus management code.
